Backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
Deprecated==1.2.13
Django==4.1.3
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==22.0
Pillow==9.3.0
pycparser==2.21
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.13
wrapt==1.14.1